+++ e2e/08-autofix/contracts/no-console/Foo1AfterFix.sol
@@ -1,0 +1,37 @@
+pragma solidity 0.8.0;
+
+
+
+
+import 'forge-std/xxxxx.sol';
+
+contract Foo1 {
+    Console[] public consoleTest;
+    Console[] public console;
+
+    struct Console {
+        uint256 one;
+        uint256 two;
+    }
+
+    function a() public {
+        
+        // comment
+        
+        uint256 declaration;
+        
+    }
+
+    function b() public {
+        
+        // comment
+        
+        uint256 declaration;
+        
+    }
+
+    function c() external {
+        consoleTest.push(0, 0);
+        console.push = (1, 1);
+    }
+}